fix: 修复消息编码中长度字段类型转换问题
将totalLength的计算和写入统一为uint类型，避免潜在的溢出和类型不匹配问题。原代码中totalLength计算为int类型，写入时强制转换为ushort可能导致数据截断，现统一使用uint确保数据完整性。

+++ GameFrameX.NetWork.Message/DefaultMessageEncoderHandler.cs
@@ -65,11 +65,11 @@ public sealed class DefaultMessageEncoderHandler : BaseMessageEncoderHandler
                 var messageBodyData = ProtoBufSerializerHelper.Serialize(messageObject);
                 byte zipFlag = 0;
                 BytesCompressHandler(ref messageBodyData, ref zipFlag);
-                var totalLength = (ushort)(PackageHeaderLength + messageBodyData.Length);
+                var totalLength = PackageHeaderLength + messageBodyData.Length;
                 var buffer = ArrayPool<byte>.Shared.Rent(totalLength);
                 var offset = 0;
                 // 总长度
-                buffer.WriteUIntValue(totalLength, ref offset);
+                buffer.WriteUIntValue((uint)totalLength, ref offset);
                 // operationType 操作类型
                 buffer.WriteByteValue((byte)messageObject.OperationType, ref offset);
                 // zipFlag 压缩标记
